+++ packages/shared/index.js
@@ -1,7 +1,0 @@
-export {
-  getWrapperStyle,
-  getPathProps,
-  getStartAt,
-  getGradientId,
-  timeStyle,
-} from './utils'