+++ packages/shared/utils/getGradientId.js
@@ -1,4 +1,0 @@
-import { uuid } from '.'
-
-export const getGradientId = (gradientUniqueKey) =>
-  `countdown-circle-timer-gradient-${gradientUniqueKey || uuid()}`