+++ packages/web/src/utils/index.js
@@ -1,3 +1,0 @@
-export { linearEase } from './linearEase'
-export { getColorsRGB, getNormalizedColors, getStroke } from './colors'
-export { visuallyHidden } from './styles'