+++ packages/web/src/utils/linearEase.js
@@ -1,8 +1,0 @@
-export const linearEase = (time, start, goal, duration) => {
-  if (duration === 0) {
-    return goal
-  }
-
-  const currentTime = time / duration
-  return start + goal * currentTime
-}